+++ src/main/java/net/yxiao233/ifeu/client/data/ModItemModelProvider.java
@@ -44,5 +44,6 @@ public class ModItemModelProvider extends ItemModelProvider {
         basicItem(ModContents.DRAGON_STAR_SWORD.get());
         basicItem(ModContents.DRAGON_STAR_PICKAXE.get());
         basicItem(ModContents.DRAGON_STAR_AXE.get());
+        basicItem(ModContents.LASER_LENS_DRAGON.get());
     }
 }