+++ src/main/java/net/yxiao233/ifeu/common/block/CreativeCapacitorBlock.java
@@ -1,0 +1,25 @@
+package net.yxiao233.ifeu.common.block;
+
+import com.buuz135.industrial.block.IndustrialBlock;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.yxiao233.ifeu.common.block.entity.CreativeCapacitorEntity;
+import net.yxiao233.ifeu.common.registry.ModItems;
+import org.jetbrains.annotations.NotNull;
+
+public class CreativeCapacitorBlock extends IndustrialBlock<CreativeCapacitorEntity> {
+    public CreativeCapacitorBlock() {
+        super("creative_capacitor", Properties.copy(Blocks.IRON_BLOCK), CreativeCapacitorEntity.class, ModItems.TAB_ADDONS);
+    }
+
+    @Override
+    public BlockEntityType.BlockEntitySupplier<?> getTileEntityFactory() {
+        return CreativeCapacitorEntity::new;
+    }
+
+    @NotNull
+    @Override
+    public RotationType getRotationType() {
+        return RotationType.FOUR_WAY;
+    }
+}