+++ src/main/java/net/yxiao233/ifeu/common/compact/jei/AbstractJEICategory.java
@@ -3,11 +3,15 @@ package net.yxiao233.ifeu.common.compact.jei;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.gui.ingredient.IRecipeSlotTooltipCallback;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.category.IRecipeCategory;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
@@ -15,6 +19,8 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Recipe;
 import net.yxiao233.ifeu.IndustrialForegoingExtraUpgrades;
+import net.yxiao233.ifeu.common.gui.AllGuiTextures;
+import net.yxiao233.ifeu.common.utils.TooltipCallBackHelper;
 import org.jetbrains.annotations.Nullable;
 
 public abstract class AbstractJEICategory<T extends Recipe<?>> implements IRecipeCategory<T> {
@@ -49,7 +55,30 @@ public abstract class AbstractJEICategory<T extends Recipe<?>> implements IRecip
     }
     @Override
     public abstract void setRecipe(IRecipeLayoutBuilder iRecipeLayoutBuilder, T t, IFocusGroup iFocusGroup);
-
     @Override
     public abstract void draw(T recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY);
+
+    public void drawTextureWithTooltip(GuiGraphics guiGraphics, AllGuiTextures allGuiTextures, Component component, int x, int y , double mouseX, double mouseY){
+        Font font = Minecraft.getInstance().font;
+
+        allGuiTextures.render(guiGraphics,x,y);
+
+        if(mouseX >= x && mouseY >= y && mouseX <= allGuiTextures.width + x && mouseY <= allGuiTextures.height + y) {
+            guiGraphics.renderTooltip(font, component, (int) mouseX, (int) mouseY);
+        }
+    }
+
+    public IRecipeSlotTooltipCallback addText(String translatableKey, ChatFormatting style){
+        return (view, tooltip) ->{
+            tooltip.add(1,Component.translatable(translatableKey).withStyle(style));
+        };
+    }
+
+    public IRecipeSlotTooltipCallback addText(TooltipCallBackHelper... tooltips){
+        return (view, tooltip) ->{
+            for (int i = 0; i < tooltips.length; i++) {
+                tooltip.add(tooltips[i].getComponent());
+            }
+        };
+    }
 }