+++ src/main/java/net/yxiao233/ifeu/common/compact/kubejs/ModKubeJSPlugin.java
@@ -20,7 +20,8 @@ public class ModKubeJSPlugin extends KubeJSPlugin {
     @Override
     public void registerRecipeSchemas(RegisterRecipeSchemasEvent event) {
         event.namespace("ifeu")
-                .register("infuser", InfuserSchema.SCHEMA);
+                .register("infuser", InfuserSchema.SCHEMA)
+                .register("arcane_dragon_egg_forging",ArcaneDragonEggForgingSchema.SCHEMA);
 
         event.namespace("industrialforegoing")
                 .register("crusher", CrusherSchema.SCHEMA)