+++ src/main/java/net/yxiao233/ifeu/common/registry/ModBlocks.java
@@ -7,6 +7,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.registries.RegistryObject;
 import net.yxiao233.ifeu.common.block.ArcaneDragonEggForgingBlock;
+import net.yxiao233.ifeu.common.block.CreativeCapacitorBlock;
 import net.yxiao233.ifeu.common.block.DragonStarGeneratorBlock;
 import net.yxiao233.ifeu.common.block.InfuserBlock;
 import org.apache.commons.lang3.tuple.Pair;
@@ -17,11 +18,13 @@ public class ModBlocks implements IModule {
     public static Pair<RegistryObject<Block>, RegistryObject<BlockEntityType<?>>> INFUSER;
     public static Pair<RegistryObject<Block>, RegistryObject<BlockEntityType<?>>> DRAGON_STAR_GENERATOR;
     public static Pair<RegistryObject<Block>, RegistryObject<BlockEntityType<?>>> ARCANE_DRAGON_EGG_FORGING;
+    public static Pair<RegistryObject<Block>, RegistryObject<BlockEntityType<?>>> CREATIVE_CAPACITOR;
     @Override
     public void generateFeatures(DeferredRegistryHelper helper) {
         INFUSER = helper.registerBlockWithTileItem("infuser", InfuserBlock::new,(blockRegistryObject) -> () -> new IndustrialBlockItem(blockRegistryObject.get(), ModItems.TAB_ADDONS), null);
         DRAGON_STAR_GENERATOR = helper.registerBlockWithTileItem("dragon_star_generator", DragonStarGeneratorBlock::new,(blockRegistryObject) -> () -> new IndustrialBlockItem(blockRegistryObject.get(), ModItems.TAB_ADDONS),null);
         ARCANE_DRAGON_EGG_FORGING = helper.registerBlockWithTileItem("arcane_dragon_egg_forging", ArcaneDragonEggForgingBlock::new,(blockRegistryObject) -> () -> new IndustrialBlockItem(blockRegistryObject.get(),ModItems.TAB_ADDONS),null);
+        CREATIVE_CAPACITOR = helper.registerBlockWithTileItem("creative_capacitor", CreativeCapacitorBlock::new,(blockRegistryObject) -> () -> new IndustrialBlockItem(blockRegistryObject.get(),ModItems.TAB_ADDONS),null);
     }
 
     public static ArrayList<RegistryObject<Block>> addBlocksToCreativeModeTab(){
@@ -30,6 +33,7 @@ public class ModBlocks implements IModule {
         blocks.add(INFUSER.getLeft());
         blocks.add(DRAGON_STAR_GENERATOR.getLeft());
         blocks.add(ARCANE_DRAGON_EGG_FORGING.getLeft());
+        blocks.add(CREATIVE_CAPACITOR.getLeft());
 
         return blocks;
     }